+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added the maxAnnotationHeight option to limit the size of annotations when using the scaled option for annotationHeight
 - Added the fontSize option for bedlike tracks
 - colorAnnotations are independent of valueScale
+- Vertical zooming on BedLike annotations
 - Added support for aggregation of `horizontal-multivec` rows by passing arrays of row indices to the `selectRows` option. With this addition also comes the track options `selectRowsAggregationMode` (for specifying an aggregation function, "mean", "sum", "variance", "deviation") and `selectRowsAggregationWithRelativeHeight` (boolean) to determine whether rows representing groups of indices have 1-unit heights or `group.length`-unit heights.
 
 _[Detailed changes since v1.9.0](https://github.com/higlass/higlass/compare/v1.9.0...develop)_